backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
langchain==0.1.0
langchain-openai==0.0.5
openai==1.6.1
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# MongoDB
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# File handling
aiofiles==23.2.1
